==> R/01-clean-data.R <==

# original source files were cleaned by anonymizing, standardizing units, and converting the 
# time variable to a numeric date representing the number of days before or after the change in 
# orthophosphate dose

# for solubility predictions, pH, alkalinity, dissolved inorganic carbon, and orthophosphate dose 
# were extracted from the source files
